Add files via upload
Purpose:  Read in 2 positive integral numbers with up to MAX_DIGITS digits, add them together
and display the result.

Note: If result is more than MAX_DIGITS, an INCORRECT result will be displayed. Therefore
if reading in 2 numbers with n digits, and the result may be n+1 digits, MAX_DIGITS should
be set to n+1 to ensure the correct result will be calculated and displayed.